refactor: 처리관련 코드 router -> controller 이동
