+++ study/yiy/13-refactoring/server/router/tweets.js
@@ -1,53 +1,22 @@
 import express from "express";
 import "express-async-errors";
-import * as tweetRepository from "../data/tweet.js";
+import * as tweetController from "../controller/tweet.js";
 const router = express.Router();
 
 // GET /tweets
 // GET /tweets?username=:username
-router.get("/", (req, res, next) => {
-  const username = req.query.username; // 없다면 undefined
-  const data = username
-    ? tweetRepository.getAllByUsername(username)
-    : tweetRepository.getAll();
-  res.status(200).json(data);
-});
+router.get("/", tweetController.getTweets);
 
 // GET /tweets/:id
-router.get("/:id", (req, res, next) => {
-  const id = req.params.id;
-  const tweet = tweetRepository.getById(id);
-  if (tweet) {
-    res.status(200).json(tweet);
-  } else {
-    return res.status(404).json({ message: `Tweet id(${id}) not found` }); // sendStatus(404)는 json 보내지지 않음.
-  }
-});
+router.get("/:id", tweetController.getTweet);
 
 // POST /tweets
-router.post("/", (req, res, next) => {
-  const { text, name, username } = req.body;
-  const tweet = tweetRepository.create(text, name, username);
-  res.status(201).json(tweet);
-});
+router.post("/", tweetController.createTweet);
 
 // PUT /tweets/:id
-router.put("/:id", (req, res, next) => {
-  const id = req.params.id;
-  const text = req.body.text;
-  const tweet = tweetRepository.update(id, text);
-  if (tweet) {
-    res.status(200).json(tweet);
-  } else {
-    res.status(404).json({ message: `Tweet id(${id}) not found` });
-  }
-});
+router.put("/:id", tweetController.updateTweet);
 
 // DELETE /tweets/:id
-router.delete("/:id", (req, res, next) => {
-  const id = req.params.id;
-  tweetRepository.remove(id);
-  res.sendStatus(204);
-});
+router.delete("/:id", tweetController.deleteTweet);
 
 export default router;
